Sun: 64 bit + CC 5.3 support

+++ cpp/demo/Freeze/phonebook/Parser.cpp
@@ -396,7 +396,7 @@ Parser::getInput(char* buf, int& result, int maxSize)
 	    }
 	}
 	
-	result = line.length();
+	result = static_cast<int>(line.length());
 	if(result > maxSize)
 	{
 	    error("input line too long");
@@ -412,7 +412,7 @@ Parser::getInput(char* buf, int& result, int maxSize)
     }
     else
     {
-	if(((result = fread(buf, 1, maxSize, yyin)) == 0) && ferror(yyin))
+	if(((result = static_cast<int>(fread(buf, 1, maxSize, yyin))) == 0) && ferror(yyin))
 	{
 	    error("input in flex scanner failed");
 	    buf[0] = EOF;
